Adds LiDAR integration.
First, adapts the lidar to run on Windows 10 as well as Linux/Mac.
Second, starts the motor running and adds an exception statement when Ctr-C is hit to stop the lidar and disconnect. Also adds a few new comments. This does not actually take any measurements as of right now.