Distance from LiDAR to Cargo!
Uses the law of cosign and the law if sign to find the distance between LiDAR and the middle of the two points we get from the hatch panel slot area.  Math is completely correct, checked by multiple sources. Code has not been fully run (yet). Should work like a charm.